10-2 add test_list controller(pro)
