--- backend/testhelper/src/main/java/kr/ac/ajou/da/testhelper/mapper/TestsMapper.java
@@ -1,13 +1,0 @@
-package kr.ac.ajou.da.testhelper.mapper;
-
-import java.util.HashMap;
-import java.util.List;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.springframework.stereotype.Repository;
-
-@Mapper
-@Repository
-public interface TestsMapper {
-	List<HashMap<String, Object>> selectTests();
-}
