--- backend/testhelper/src/main/resources/mapper/test.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="kr.ac.ajou.da.testhelper.mapper.TestsMapper">
-	<select id="selectTests" resultType="hashMap">
-		select * from test
-	</select>
-</mapper>
